requirements.txt
PyYAML>=6.0,<7
PyQt6>=6.5,<7
